Removed '//swiftlint:disable all' rule
