+++ CodeEdit/Info.plist
@@ -45,6 +45,6 @@
 		</dict>
 	</array>
 	<key>GitHash</key>
-	<string>f05d0e52c02a3cd7e2db677f015ee09ef7ed467f</string>
+	<string>74affa37b12d7b190a89101fde28f3b26f4c09a2</string>
 </dict>
 </plist>
